Clean up database dumping process
Previous state of database dumping was confusing to me. I worry there might be
a good motivation to keep old dumps around, although it seems that a lot of
old dumps around are unused. I removed all the unused dumps, keeping only
bi-directional json mapping from and to new ids.

The old way of db dumping used taxonomy_to_concept_v67.json file as a source of
truth for some ids that were dumped from datomic. I removed it, since data in
the newest database to get dumps from (jobtech-taxonomy-development-2020-03-18-2)
is matching the data in v67.

Previous dumping process also renamed all types to kebab-case, which is
effectively noop on current types, so I removed that as well: it seems to be a
relic of some old schema.